requirements.txt
jupyterlab
notebook
streamlit
ibm-watsonx-ai
opencv-python
openmeteo_requests
requests-cache
retry-requests
numpy
pandas
python-dotenv
geopandas
scikit-learn
seaborn
supabase
streamlit_js_eval
geopy